allow automatic repl reload if ob-scala-cli block params changed
Also support scala-version.

The problem was that evaluating two blocks with different dependencies in
succession would make the second block not work,

 #+begin-src scala :dep '("a")
 #+end src

 #+begin-src scala :dep '("b")
 // this wouldn't reload scala-repl with the right dependency
 #+end src

because scala-cli was
loaded with the first dependencies.

This enforces an automatic reload of scala-cli in that case.

The :session is not supported yet, but if you want to emulate just put
all the dependencies in the first block (so the repl will not be
restarted).