improve parsing and fix process-filter
Setting the process-filter to nil would break the
emulation (management of color characters) given by
term-emulate-terminal, so I have changed to reset to that function.

Also, it seems that the file name passed to the parsing functions is
not always correct, so I am now splitting on repl.sc

for instance I had

```
file:
/tmp/ob-scala-cli/8365dfae-afef-416b-9a96-c8536d8515cc/repl.sc

result first line:
:load /tmp/ob-scala-cli/8365d8365dfae-afef-416b-9a96-c8536d8515cc/repl.sc
```

which I couldn't figure out how it is a problem on our side: I have a
feeling that it may be a scala-cli problem, but I didn't dig too much.